+++ examples/basic-iframe/pages/posts/[slug].js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { staticRequest } from 'tinacms'
 import { Layout } from '../../components/Layout'
-import { useTina } from 'tinacms/dist/react'
+import { useEditState, useTina } from 'tinacms/dist/react'
 
 const query = `query getPost($relativePath: String!) {
   post(relativePath: $relativePath) {
@@ -29,6 +29,8 @@ export default function Home(props) {
     variables: props.variables,
     data: props.data,
   })
+  const { edit } = useEditState()
+  console.log('edit', edit)
 
   const cleanedObject = React.useMemo(() => {
     const obj = {}