+++ experimental-examples/tina-cloud-starter-iframe/.tina/schema.ts
@@ -11,7 +11,6 @@ const tem = {
   label: "test",
   fields: [{ type: "string", name: "yo" }],
 };
-// const CreateInnerTemplate = () => {
 const InnerTemplate: TinaTemplate = {
   label: "List Items",
   name: "items",
@@ -36,8 +35,6 @@ const InnerTemplate: TinaTemplate = {
     },
   ],
 };
-// return InnerTemplate;
-// };
 
 const BlocksTemplate: TinaTemplate = {
   label: "Page Blocks",
@@ -233,6 +230,10 @@ const schema = defineSchema({
       path: "content/global",
       ui: {
         global: true,
+        allowedActions: {
+          create: false,
+          delete: false,
+        },
       },
       format: "json",
       fields: [