+++ packages/@tinacms/schema-tools/src/types/SchemaTypes.ts
@@ -39,6 +39,9 @@ export type UIField<F extends UIField = any, Shape = any> = {
 export interface TinaCloudSchema<WithNamespace extends boolean, Store = any> {
   templates?: GlobalTemplate<WithNamespace>[]
   collections: TinaCloudCollection<WithNamespace>[]
+  /**
+   * @deprecated use `defineConfig` in a config.{js,ts} file instead
+   */
   config?: TinaCloudSchemaConfig<Store>
 }
 export type TinaCloudSchemaBase = TinaCloudSchema<false>
@@ -104,6 +107,12 @@ export interface UICollection {
      */
     readonly?: boolean
   }
+
+  allowedActions?: {
+    create?: boolean
+    delete?: boolean
+  }
+
   /**
    * Forms for this collection will be editable from the global sidebar rather than the form panel
    */