+++ packages/@tinacms/schema-tools/src/validate/tinaCloudSchemaConfig.ts
@@ -11,7 +11,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import { TinaCloudSchemaConfig } from '../types'
+import { TinaCloudSchemaConfig } from '../types/index'
 
 import z from 'zod'
 const tinaConfigKey = z