+++ packages/next-tinacms-cloudinary/src/cloudinary-media-store.ts
@@ -25,7 +25,7 @@ export class CloudinaryMediaStore implements MediaStore {
   fetchFunction = (input: RequestInfo, init?: RequestInit) => {
     return fetch(input, init)
   }
-  accept = 'text/*,  application/*, image/*'
+  accept = 'text/*,application/*,image/*'
 
   async persist(media: MediaUploadOptions[]): Promise<Media[]> {
     let newFiles: Media[] = []
@@ -66,6 +66,7 @@ export class CloudinaryMediaStore implements MediaStore {
         filename: fileRes.original_filename,
         directory: '/',
         previewSrc: fileRes.url,
+        src: fileRes.url,
       }
 
       newFiles.push(parsedRes)
@@ -109,7 +110,7 @@ export class CloudinaryMediaStore implements MediaStore {
   }
 
   parse = (img) => {
-    return img.previewSrc
+    return img.src
   }
 
   private buildQuery(options: MediaListOptions) {