+++ experimental-examples/unit-test-example/CHANGELOG.md
@@ -4,6 +4,40 @@
 
 ### Patch Changes
 
+- Updated dependencies [958d10c82]
+  - tinacms@1.0.0
+
+## null
+
+### Patch Changes
+
+- Updated dependencies [a5d6722c7]
+  - tinacms@0.70.2
+
+## null
+
+### Patch Changes
+
+- Updated dependencies [6c93834a2]
+- Updated dependencies [b12957885]
+  - tinacms@0.70.1
+
+## null
+
+### Patch Changes
+
+- Updated dependencies [2e0a98d61]
+- Updated dependencies [f7d3cf532]
+- Updated dependencies [c3f307f46]
+- Updated dependencies [774abcf9c]
+- Updated dependencies [0ead00fa8]
+- Updated dependencies [8f28bfbd5]
+  - tinacms@0.70.0
+
+## null
+
+### Patch Changes
+
 - tinacms@0.69.22
 
 ## null