+++ packages/@tinacms/graphql/CHANGELOG.md
@@ -1,5 +1,47 @@
 # tina-graphql
 
+## 1.0.0
+
+### Major Changes
+
+- 958d10c82: Tina 1.0 Release
+
+  Make sure you have updated to th "iframe" path: https://tina.io/blog/upgrading-to-iframe/
+
+### Patch Changes
+
+- Updated dependencies [958d10c82]
+  - @tinacms/datalayer@1.0.0
+  - @tinacms/mdx@1.0.0
+  - @tinacms/schema-tools@1.0.0
+
+## 0.63.20
+
+### Patch Changes
+
+- Updated dependencies [a5d6722c7]
+- Updated dependencies [14c5cdffe]
+  - @tinacms/schema-tools@0.2.2
+  - @tinacms/mdx@0.61.17
+
+## 0.63.19
+
+### Patch Changes
+
+- Updated dependencies [6c93834a2]
+- Updated dependencies [4b174e14b]
+  - @tinacms/schema-tools@0.2.1
+  - @tinacms/mdx@0.61.16
+
+## 0.63.18
+
+### Patch Changes
+
+- Updated dependencies [774abcf9c]
+- Updated dependencies [245a65dfe]
+  - @tinacms/schema-tools@0.2.0
+  - @tinacms/mdx@0.61.15
+
 ## 0.63.17
 
 ### Patch Changes