+++ packages/@tinacms/toolkit/CHANGELOG.md
@@ -1,5 +1,24 @@
 # Change Log
 
+## 1.0.0
+
+### Major Changes
+
+- 958d10c82: Tina 1.0 Release
+
+  Make sure you have updated to th "iframe" path: https://tina.io/blog/upgrading-to-iframe/
+
+### Patch Changes
+
+- Updated dependencies [958d10c82]
+  - @tinacms/sharedctx@1.0.0
+
+## 0.58.5
+
+### Patch Changes
+
+- 7dda334e5: fix wrong background when hovering the "Embed" button inside the toolbar
+
 ## 0.58.4
 
 ### Patch Changes