+++ packages/next-tinacms-cloudinary/CHANGELOG.md
@@ -1,5 +1,19 @@
 # next-tinacms-cloudinary
 
+## 4.0.0
+
+### Major Changes
+
+- 958d10c82: Tina 1.0 Release
+
+  Make sure you have updated to th "iframe" path: https://tina.io/blog/upgrading-to-iframe/
+
+## 3.5.35
+
+### Patch Changes
+
+- 3370cefd8: Fix preview images in cloudinary
+
 ## 3.5.34
 
 ### Patch Changes